Move accountstorage to internal/interface
